prod-con/producer/requirements_producer.txt
psycopg2-binary
sqlalchemy
sqlalchemy.orm
spotipy
requests
confluent-kafka
python-dotenv